Avances en chart

+++ Cresud.CDP.Dtos/CartaDePorteGraficoItem.cs
@@ -4,7 +4,7 @@ namespace Cresud.CDP.Dtos
 {
     public class CartaDePorteGraficoItem
     {
-        public int EmpresaId { get; set; }
+        public long EmpresaId { get; set; }
         public DateTime Fecha { get; set; }
         public int CantidadAfip { get; set; }
         public int CantidadSap { get; set; }
